requirements.txt
Adafruit-Blinka==6.18.0
adafruit-circuitpython-dht==3.7.0
Adafruit-DHT==1.4.0
Adafruit-PlatformDetect==3.19.2
Adafruit-PureIO==1.1.9
aiohttp==3.8.1
aiosignal==1.2.0
ansicon==1.89.0
asgiref==3.4.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
autopep8==1.6.0
board==1.0
certifi==2021.10.8
charset-normalizer==2.0.9
Django==3.2.10
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
frozenlist==1.2.0
idna==3.3
jinxed==1.1.0
multidict==5.2.0
pkg-resources==0.0.0
pycodestyle==2.8.0
pyftdi==0.53.3
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pyusb==1.2.1
requests==2.27.0
rpi-ws281x==4.3.1
RPi.GPIO==0.7.0
schedule==1.1.0
six==1.16.0
sqlparse==0.4.2
sysv-ipc==1.1.0
telepot==12.7
toml==0.10.2
typing-extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
yarl==1.7.2
